Move all mapping code to 'mappers' directory
