+++ package.json
@@ -65,6 +65,7 @@
     "firstline": "^2.0.2",
     "good-winston": "^4.0.0",
     "hapi-auth-jwt2": "^8.8.1",
+    "install": "^0.13.0",
     "knex": "^0.20.8",
     "libxmljs": "^0.19.7",
     "lodash": "^4.17.15",
@@ -72,6 +73,7 @@
     "moment-range": "^4.0.2",
     "node-cron": "^1.2.1",
     "notifications-node-client": "^4.7.0",
+    "npm": "^6.13.6",
     "nunjucks": "^3.2.0",
     "path": "^0.12.7",
     "pg": "^7.17.1",
