+++ src/modules/billing/mappers/index.js
@@ -1,0 +1,13 @@
+exports.address = require('./address');
+exports.abstractionPeriod = require('./abstraction-period');
+exports.batch = require('./batch');
+exports.agreements = require('./agreements');
+exports.chargeElements = require('./charge-elements');
+exports.company = require('./company');
+exports.contact = require('./contact');
+exports.invoice = require('./invoice');
+exports.invoiceAccount = require('./invoice-account');
+exports.invoiceLicence = require('./invoice-licence');
+exports.licence = require('./licence');
+exports.region = require('./region');
+exports.transactions = require('./transactions');
