+++ src/modules/billing/mappers/region.js
@@ -1,0 +1,38 @@
+'use strict';
+
+const Region = require('../../../lib/models/region');
+
+/**
+ * Maps a row from water.regions to a Region model
+ * @param {Object} row - camel cased
+ * @return {Region}
+ */
+const dbToModel = row => {
+  const region = new Region();
+  region.fromHash({
+    type: Region.types.region,
+    id: row.regionId,
+    name: row.name,
+    code: row.chargeRegionId,
+    numericCode: row.naldRegionId
+  });
+  return region;
+};
+
+const create = (type, props = {}) => {
+  const region = new Region();
+  region.fromHash(props);
+  return region;
+};
+
+/**
+ * @param {Object} regions - from water.licences.regions jsonb field
+ * @return {Object} region instances
+ */
+const licenceRegionsToModels = regions => ({
+  historicalArea: create(Region.types.environmentAgencyArea, { code: regions.historicalAreaCode }),
+  regionalChargeArea: create(Region.types.regionalChargeArea, { name: regions.regionalChargeArea })
+});
+
+exports.dbToModel = dbToModel;
+exports.licenceRegionsToModels = licenceRegionsToModels;
