+++ src/modules/billing/services/batch-service.js
@@ -1,23 +1,6 @@
-const Batch = require('../../../lib/models/batch');
-const Region = require('../../../lib/models/region');
 const repos = require('../../../lib/connectors/repository');
-const FinancialYear = require('../../../lib/models/financial-year');
-const regionsService = require('./regions-service');
-
-const mapDBToModel = row => {
-  const batch = new Batch();
-  batch.fromHash({
-    id: row.billing_batch_id,
-    type: row.batch_type,
-    season: row.season,
-    startYear: new FinancialYear(row.from_financial_year_ending),
-    endYear: new FinancialYear(row.to_financial_year_ending),
-    status: row.status,
-    dateCreated: row.date_created,
-    region: regionsService.mapDBToModel(row.region)
-  });
-  return batch;
-};
+const newRepos = require('../../../lib/connectors/repos');
+const mappers = require('../mappers');
 
 /**
  * Loads a Batch instance by ID
@@ -25,14 +8,14 @@ const mapDBToModel = row => {
  * @return {Promise<Batch>}
  */
 const getBatchById = async id => {
-  const row = await repos.billingBatches.getById(id);
-  return mapDBToModel(row);
+  const row = await newRepos.billingBatches.findOne(id);
+  return mappers.batch.dbToModel(row);
 };
 
 const getBatches = async (page = 1, perPage = Number.MAX_SAFE_INTEGER) => {
   const pagination = { page, perPage };
   const { rows } = await repos.billingBatches.find(null, { date_created: -1 }, pagination);
-  const batches = rows.map(mapDBToModel);
+  const batches = rows.map(mappers.batch.dbToModel);
 
   const { rows: [totalRowCount] } = await repos.billingBatches.findRowCount();
   const totalRows = totalRowCount.totalrowcount;
@@ -47,6 +30,5 @@ const getBatches = async (page = 1, perPage = Number.MAX_SAFE_INTEGER) => {
   };
 };
 
-exports.mapDBToModel = mapDBToModel;
 exports.getBatchById = getBatchById;
 exports.getBatches = getBatches;
