+++ src/modules/billing/services/invoice-accounts-service.js
@@ -1,27 +1,7 @@
 'use strict';
 
-const InvoiceAccount = require('../../../lib/models/invoice-account');
-const Company = require('../../../lib/models/company');
 const invoiceAccountsConnector = require('../../../lib/connectors/crm-v2/invoice-accounts');
-
-/**
- * Maps CRM invoice account and (optionally) company data to a water service model
- * @param {Object} invoiceAccount - CRM invoice account data
- * @param {Object} company - CRM company data
- * @return {InvoiceAccount}
- */
-const mapCRMInvoiceAccountToModel = (invoiceAccount, company) => {
-  const invoiceAccountModel = new InvoiceAccount(invoiceAccount.invoiceAccountId);
-  invoiceAccountModel.accountNumber = invoiceAccount.invoiceAccountNumber;
-
-  if (company) {
-    const companyModel = new Company(company.companyId);
-    companyModel.pickFrom(company, ['type', 'name']);
-    invoiceAccountModel.company = companyModel;
-  }
-
-  return invoiceAccountModel;
-};
+const mappers = require('../mappers');
 
 /**
  * Gets invoice accounts with specified IDs from CRM and
@@ -37,7 +17,7 @@ const getByInvoiceAccountIds = async (ids = []) => {
   const invoiceAccounts = await invoiceAccountsConnector.getInvoiceAccountsByIds(ids);
 
   return invoiceAccounts.map(invoiceAccount =>
-    mapCRMInvoiceAccountToModel(invoiceAccount, invoiceAccount.company)
+    mappers.invoiceAccount.crmToModel(invoiceAccount, invoiceAccount.company)
   );
 };
 
@@ -49,9 +29,8 @@ const getByInvoiceAccountIds = async (ids = []) => {
  */
 const getByInvoiceAccountId = async id => {
   const invoiceAccount = await invoiceAccountsConnector.getInvoiceAccountById(id);
-  return mapCRMInvoiceAccountToModel(invoiceAccount, invoiceAccount.company);
+  return mappers.invoiceAccount.crmToModel(invoiceAccount, invoiceAccount.company);
 };
 
-exports.mapCRMInvoiceAccountToModel = mapCRMInvoiceAccountToModel;
 exports.getByInvoiceAccountIds = getByInvoiceAccountIds;
 exports.getByInvoiceAccountId = getByInvoiceAccountId;
