+++ src/modules/billing/services/licence-service.js
@@ -1,41 +1,7 @@
+'use strict';
+
 const repos = require('../../../lib/connectors/repository');
-const Licence = require('../../../lib/models/licence');
-const Region = require('../../../lib/models/region');
-
-const mapDBToModel = (row) => {
-  const licence = new Licence();
-  licence.fromHash({
-    id: row.licence_id,
-    licenceNumber: row.licence_ref,
-    isWaterUndertaker: row.is_water_undertaker
-  });
-
-  // NALD region
-  licence.region = new Region();
-  licence.region.fromHash({
-    type: Region.types.region,
-    id: row.region.region_id,
-    name: row.region.name,
-    code: row.region.charge_region_id,
-    numericCode: row.region.nald_region_id
-  });
-
-  // Historical EA area
-  licence.historicalArea = new Region();
-  licence.historicalArea.fromHash({
-    type: Region.types.environmentAgencyArea,
-    code: row.regions.historicalAreaCode
-  });
-
-  // Regional charging area
-  licence.regionalChargeArea = new Region();
-  licence.regionalChargeArea.fromHash({
-    type: Region.types.regionalChargeArea,
-    name: row.regions.regionalChargeArea
-  });
-
-  return licence;
-};
+const mappers = require('../mappers');
 
 /**
  * Gets a licence by licence number
@@ -44,8 +10,7 @@ const mapDBToModel = (row) => {
  */
 const getByLicenceNumber = async licenceNumber => {
   const row = await repos.licences.findOneByLicenceNumber(licenceNumber);
-  return mapDBToModel(row);
+  return mappers.licence.dbToModel(row);
 };
 
-exports.mapDBToModel = mapDBToModel;
 exports.getByLicenceNumber = getByLicenceNumber;
